#154, refactor as kotlin files
