--- app/src/main/kotlin/org/artembogomolova/demo/webapp/main/dao/repo/IPhysicalAddressRepository.kt
@@ -1,8 +1,0 @@
-package org.artembogomolova.demo.webapp.main.dao.repo
-
-import org.artembogomolova.demo.webapp.main.domain.core.PhysicalAddress
-import org.springframework.data.repository.CrudRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface IPhysicalAddressRepository : CrudRepository<PhysicalAddress?, Long?>
